+++ packages/client/components/ReflectionCard/ReflectionCardAuthor.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled'
 import {PALETTE} from '../../styles/paletteV3'
 
 const Author = styled('div')({
-  padding: '12px 16px',
+  padding: '8px 16px',
   color: PALETTE.SLATE_600,
   fontSize: 14,
   fontWeight: 700,