+++ packages/mattermost-plugin/components/ModalRoot.tsx
@@ -1,3 +1,6 @@
+import CreateTaskModalRoot from './CreateTaskModal'
+import InviteToMeetingModalRoot from './InviteToMeetingModal'
+import InviteToTeamModalRoot from './InviteToTeamModal'
 import LinkTeamModal from './LinkTeamModal'
 import PushReflectionModalRoot from './PushReflection'
 import StartActivityModalRoot from './StartActivityModal'
@@ -5,9 +8,12 @@ import StartActivityModalRoot from './StartActivityModal'
 const ModalRoot = () => {
   return (
     <>
-      <StartActivityModalRoot />
+      <CreateTaskModalRoot />
+      <InviteToTeamModalRoot />
+      <InviteToMeetingModalRoot />
       <LinkTeamModal />
       <PushReflectionModalRoot />
+      <StartActivityModalRoot />
     </>
   )
 }