+++ packages/mattermost-plugin/reducers.ts
@@ -52,6 +52,9 @@ const localSlice = createSlice({
   name: 'local',
   initialState: {
     isStartActivityModalVisible: false,
+    isCreateTaskModalVisible: false,
+    isInviteToTeamModalVisible: false,
+    isInviteToMeetingModalVisible: false,
     pushPostAsReflection: null as string | null,
     isLinkTeamModalVisible: false,
     linkedTeamIds: {} as Record<string, {loading: boolean; teamIds: string[]}>
@@ -62,6 +65,24 @@ const localSlice = createSlice({
     },
     closeStartActivityModal: (state) => {
       state.isStartActivityModalVisible = false
+    },
+    openCreateTaskModal: (state) => {
+      state.isCreateTaskModalVisible = true
+    },
+    closeCreateTaskModal: (state) => {
+      state.isCreateTaskModalVisible = false
+    },
+    openInviteToTeamModal: (state) => {
+      state.isInviteToTeamModalVisible = true
+    },
+    closeInviteToTeamModal: (state) => {
+      state.isInviteToTeamModalVisible = false
+    },
+    openInviteToMeetingModal: (state) => {
+      state.isInviteToMeetingModalVisible = true
+    },
+    closeInviteToMeetingModal: (state) => {
+      state.isInviteToMeetingModalVisible = false
     },
     openPushPostAsReflection: (state, action: PayloadAction<string>) => {
       state.pushPostAsReflection = action.payload
@@ -117,6 +138,12 @@ const localSlice = createSlice({
 export const {
   openStartActivityModal,
   closeStartActivityModal,
+  openCreateTaskModal,
+  closeCreateTaskModal,
+  openInviteToTeamModal,
+  closeInviteToTeamModal,
+  openInviteToMeetingModal,
+  closeInviteToMeetingModal,
   openPushPostAsReflection,
   closePushPostAsReflection,
   openLinkTeamModal,