+++ packages/mattermost-plugin/selectors.ts
@@ -62,6 +62,15 @@ export const getPluginState = (state: GlobalState) =>
 export const isStartActivityModalVisible = (state: GlobalState) =>
   getPluginState(state).isStartActivityModalVisible
 
+export const isCreateTaskModalVisible = (state: GlobalState) =>
+  getPluginState(state).isCreateTaskModalVisible
+
+export const isInviteToTeamModalVisible = (state: GlobalState) =>
+  getPluginState(state).isInviteToTeamModalVisible
+
+export const isInviteToMeetingModalVisible = (state: GlobalState) =>
+  getPluginState(state).isInviteToMeetingModalVisible
+
 export const isLinkTeamModalVisible = (state: GlobalState) =>
   getPluginState(state).isLinkTeamModalVisible